[fix]pkg/view
