+++ pkg/server/view/auth.go
@@ -1,7 +1,5 @@
 package view
 
-import "github.com/gin-gonic/gin"
-
 type SignUpResponse struct{
 	Token string `json:"token"`
 	LoginId string `json:"loginID"`
@@ -13,19 +11,4 @@ func ReturnSignUpResponse(token string, uuid string) SignUpResponse {
 		LoginId: uuid,
 	}
 	return body
-}
-
-type Error struct {
-	Code		int		`json:"code"`
-	Message		string	`json:"message"`
-	Description	string	`json:"description"`
-}
-
-func ReturnErrorResponse(c *gin.Context, code int, msg, desc string) {
-	body := Error{
-		Code: code,
-		Message: msg,
-		Description: desc,
-	}
-	c.JSON(code, body)
 }
